+++ cli/src/core/package_json/package_json_constants.rs
@@ -9,7 +9,7 @@ use crate::{
 // @biomejs/biome
 pub(crate) const BIOME_VERSION: &str = "1.9.4";
 // @eslint/js
-pub(crate) const ESLINT_VERSION: &str = "^9.38.0";
+pub(crate) const ESLINT_VERSION: &str = "^9.39.1";
 // @types/jest
 pub(crate) const JEST_TYPES_VERSION: &str = "^30.0.0";
 // jest
@@ -23,19 +23,19 @@ pub(crate) const TS_JEST_VERSION: &str = "^29.4.5";
 // ts-nodef
 pub(crate) const TS_NODE_VERSION: &str = "^10.9.2";
 // typescript-eslint
-pub(crate) const TYPESCRIPT_ESLINT_VERSION: &str = "^8.46.2";
+pub(crate) const TYPESCRIPT_ESLINT_VERSION: &str = "^8.46.4";
 // vitest
-pub(crate) const VITEST_VERSION: &str = "^4.0.4";
+pub(crate) const VITEST_VERSION: &str = "^4.0.9";
 
 // Application package.json dependencies constants
 // globals
-pub(crate) const GLOBALS_VERSION: &str = "^16.4.0";
+pub(crate) const GLOBALS_VERSION: &str = "^16.5.0";
 // husky
 pub(crate) const HUSKY_VERSION: &str = "^9.1.7";
 // lint-staged
 pub(crate) const LINT_STAGED_VERSION: &str = "^15.4.3";
 // node-gyp
-pub(crate) const NODE_GYP_VERSION: &str = "^11.5.0";
+pub(crate) const NODE_GYP_VERSION: &str = "^12.1.0";
 // sort-package-json
 pub(crate) const SORT_PACKAGE_JSON_VERSION: &str = "^3.0.0";
 // tsx
@@ -219,55 +219,55 @@ pub(crate) const APP_UNIVERSAL_SDK_VERSION: &str = "workspace:*";
 // @forklaunch/bunrun
 pub(crate) const BUNRUN_VERSION: &str = "~0.0.2";
 // @forklaunch/common
-pub(crate) const COMMON_VERSION: &str = "~0.6.22";
+pub(crate) const COMMON_VERSION: &str = "~0.6.23";
 // @forklaunch/core
-pub(crate) const CORE_VERSION: &str = "~0.15.12";
+pub(crate) const CORE_VERSION: &str = "~0.16.1";
 // @forklaunch/express
-pub(crate) const EXPRESS_VERSION: &str = "~0.9.31";
+pub(crate) const EXPRESS_VERSION: &str = "~0.9.33";
 // @forklaunch/hyper-express
-pub(crate) const HYPER_EXPRESS_VERSION: &str = "~0.9.31";
+pub(crate) const HYPER_EXPRESS_VERSION: &str = "~0.10.2";
 // @forklaunch/implementation-billing-base
-pub(crate) const BILLING_BASE_VERSION: &str = "~0.8.6";
+pub(crate) const BILLING_BASE_VERSION: &str = "~0.8.7";
 // @forklaunch/implementation-billing-stripe
-pub(crate) const BILLING_STRIPE_VERSION: &str = "~0.5.6";
+pub(crate) const BILLING_STRIPE_VERSION: &str = "~0.5.7";
 // @forklaunch/implementation-iam-base
-pub(crate) const IAM_BASE_VERSION: &str = "~0.8.6";
+pub(crate) const IAM_BASE_VERSION: &str = "~0.8.7";
 // @forklaunch/implementation-worker-bullmq
-pub(crate) const WORKER_BULLMQ_VERSION: &str = "~0.8.6";
+pub(crate) const WORKER_BULLMQ_VERSION: &str = "~0.8.7";
 // @forklaunch/implementation-worker-redis
-pub(crate) const WORKER_REDIS_VERSION: &str = "~0.8.6";
+pub(crate) const WORKER_REDIS_VERSION: &str = "~0.8.7";
 // @forklaunch/implementation-worker-database
-pub(crate) const WORKER_DATABASE_VERSION: &str = "~0.8.6";
+pub(crate) const WORKER_DATABASE_VERSION: &str = "~0.8.7";
 // @forklaunch/implementation-worker-kafka
-pub(crate) const WORKER_KAFKA_VERSION: &str = "~0.9.0";
+pub(crate) const WORKER_KAFKA_VERSION: &str = "~0.9.1";
 // @forklaunch/infrastructure-redis
-pub(crate) const INFRASTRUCTURE_REDIS_VERSION: &str = "~0.4.30";
+pub(crate) const INFRASTRUCTURE_REDIS_VERSION: &str = "~0.4.32";
 // @forklaunch/infrastructure-s3
-pub(crate) const INFRASTRUCTURE_S3_VERSION: &str = "~0.4.30";
+pub(crate) const INFRASTRUCTURE_S3_VERSION: &str = "~0.4.32";
 // @forklaunch/interfaces-billing
-pub(crate) const BILLING_INTERFACES_VERSION: &str = "~0.8.6";
+pub(crate) const BILLING_INTERFACES_VERSION: &str = "~0.8.7";
 // @forklaunch/interfaces-iam
-pub(crate) const IAM_INTERFACES_VERSION: &str = "~0.8.6";
+pub(crate) const IAM_INTERFACES_VERSION: &str = "~0.8.7";
 // @forklaunch/interfaces-worker
-pub(crate) const WORKER_INTERFACES_VERSION: &str = "~0.7.6";
+pub(crate) const WORKER_INTERFACES_VERSION: &str = "~0.7.7";
 // @forklaunch/internal
-pub(crate) const INTERNAL_VERSION: &str = "~0.3.22";
+pub(crate) const INTERNAL_VERSION: &str = "~0.3.23";
 // @forklaunch/testing
-pub(crate) const TESTING_VERSION: &str = "~0.0.11";
+pub(crate) const TESTING_VERSION: &str = "~0.0.12";
 // @forklaunch/universal-sdk
-pub(crate) const UNIVERSAL_SDK_VERSION: &str = "~0.7.22";
+pub(crate) const UNIVERSAL_SDK_VERSION: &str = "~0.7.23";
 // @forklaunch/validator
-pub(crate) const VALIDATOR_VERSION: &str = "~0.10.22";
+pub(crate) const VALIDATOR_VERSION: &str = "~0.10.23";
 // @mikro-orm/core
-pub(crate) const MIKRO_ORM_CORE_VERSION: &str = "^6.5.9";
+pub(crate) const MIKRO_ORM_CORE_VERSION: &str = "^6.6.0";
 // @mikro-orm/migrations
-pub(crate) const MIKRO_ORM_MIGRATIONS_VERSION: &str = "^6.5.9";
+pub(crate) const MIKRO_ORM_MIGRATIONS_VERSION: &str = "^6.6.0";
 // @mikro-orm/postgresql,@mikro-orm/mongodb,@mikro-orm/mysql,@mikro-orm/better-sqlite,@mikro-orm/sqlite,@mikro-orm/mariadb,@mikro-orm/libsql,@mikro-orm/mssql
-pub(crate) const MIKRO_ORM_DATABASE_VERSION: &str = "^6.5.9";
+pub(crate) const MIKRO_ORM_DATABASE_VERSION: &str = "^6.6.0";
 // @mikro-orm/reflection
-pub(crate) const MIKRO_ORM_REFLECTION_VERSION: &str = "^6.5.9";
+pub(crate) const MIKRO_ORM_REFLECTION_VERSION: &str = "^6.6.0";
 // @mikro-orm/seeder
-pub(crate) const MIKRO_ORM_SEEDER_VERSION: &str = "^6.5.9";
+pub(crate) const MIKRO_ORM_SEEDER_VERSION: &str = "^6.6.0";
 // @opentelemetry/api
 pub(crate) const OPENTELEMETRY_API_VERSION: &str = "^1.9.0";
 // @sinclair/typebox
@@ -275,19 +275,19 @@ pub(crate) const TYPEBOX_VERSION: &str = "^0.34.41";
 // ajv
 pub(crate) const AJV_VERSION: &str = "^8.17.1";
 // better-auth
-pub(crate) const BETTER_AUTH_VERSION: &str = "^1.3.33";
+pub(crate) const BETTER_AUTH_VERSION: &str = "^1.3.34";
 // bullmq
-pub(crate) const BULLMQ_VERSION: &str = "^5.62.0";
+pub(crate) const BULLMQ_VERSION: &str = "^5.63.2";
 // better-sqlite3
 pub(crate) const BETTER_SQLITE3_VERSION: &str = "^12.4.1";
 // dotenv
 pub(crate) const DOTENV_VERSION: &str = "^17.2.3";
 // jose
-pub(crate) const JOSE_VERSION: &str = "^6.1.0";
+pub(crate) const JOSE_VERSION: &str = "^6.1.2";
 // sqlite3
 pub(crate) const SQLITE3_VERSION: &str = "^5.1.7";
 // stripe
-pub(crate) const STRIPE_VERSION: &str = "^19.1.0";
+pub(crate) const STRIPE_VERSION: &str = "^19.3.1";
 // uuid
 pub(crate) const UUID_VERSION: &str = "^13.0.0";
 // zod
@@ -295,7 +295,7 @@ pub(crate) const ZOD_VERSION: &str = "^4.1.12";
 
 // Project package.json devDependencies constants
 // @mikro-orm/cli
-pub(crate) const MIKRO_ORM_CLI_VERSION: &str = "^6.5.9";
+pub(crate) const MIKRO_ORM_CLI_VERSION: &str = "^6.6.0";
 // @types/express
 pub(crate) const TYPES_EXPRESS_VERSION: &str = "^5.0.5";
 // @types/express-serve-static-core