+++ blueprint/implementations/iam/base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/implementation-iam-base
 
+## 0.8.7
+
+### Patch Changes
+
+- Update package versions
+- Updated dependencies
+  - @forklaunch/interfaces-iam@0.8.7
+
 ## 0.8.6
 
 ### Patch Changes