+++ framework/ws/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @forklaunch/ws
 
+## 0.1.2
+
+### Patch Changes
+
+- update framework pages
+- Updated dependencies
+  - @forklaunch/validator@0.10.23
+  - @forklaunch/common@0.6.23
+  - @forklaunch/core@0.16.1
+
 ## 0.1.1
 
 ### Patch Changes